Simplify pip install step using ensurepip
Use ensurepip instead of fetching get-pip to install pip, to make the build
step simpler. Note that --without-ensurepip in the Python configure args is
still desired and correct, as it only prevents from running ensurepip after the
build. The ensurepip and upgrade to the pinned version should be run in the
same step to prevent layer bloat by modified files.