Reduce number of layers and optimize build of the base-python images (#321)
* Bind-mount Python patches instead of copying them to the image

We don't need the patches in runtime but because they're copied, they are still
persisted in a layer of the image. Use bind-mount to have them available only
for the build.

* Create symlinks in the Python build step

There is no major benefit in splitting these two actions to two steps.

* Define PIP_VERSION arg once it's needed

By defining it early we're busting the cache on pip version upgrade,
effectively making the split into two layers pointless.

* Simplify pip install step using ensurepip

Use ensurepip instead of fetching get-pip to install pip, to make the build
step simpler. Note that --without-ensurepip in the Python configure args is
still desired and correct, as it only prevents from running ensurepip after the
build. The ensurepip and upgrade to the pinned version should be run in the
same step to prevent layer bloat by modified files.

* Use --default-pip instead of creating a symlink