Drop deprecated architectures, build aarch64 on ARM runners (#324)
* Drop build of deprecated architectures

Do not build base images for deprecated architectures. This means removal of
build for all armv7, armhf and i386 image variants. Since there is no Raspbian
image for any of the supported architectures, these base images are removed
completely.

* Remove deprecated architectures from Dockerfile comments and S6_ARCH handling

* Use arm64 runners for aarch64 builds

As we're now effectively building only aarch64 or amd64, we can simply pass the
arch to the builder image argument.

* Remove unnecessary matrix excludes

* Use matrix include pattern to select builder as suggested in code review